Reconfigure build config and scripts
The idea with this change is to have the 'module' entry in `package.json` be what's emitted with the typescript compiler and config. The intent here is to remove the additional level of indirection that bundling creates with the source files. The rationale being the module entry is for ES module aware consumers, in that case the distributed source files should be as close to the canonical source files as possible. The requirement of the consumer being that it's ESM aware, and resolves packages node style. I believe Webpack and Rollup can be easily configured to match this, if it's not already out of the box.
In contrast to this, and for convenience and universal usage, the 'main' entry UMD bundle is the result of what *is* processed through rollup.